+++ alphadia/outputtransform_/outputtransform.py
@@ -9,20 +9,18 @@ import pandas as pd
 from alphabase.peptide import fragment, precursor
 from alphabase.spectral_library import base
 from alphabase.spectral_library.base import SpecLibBase
-from outputtransform_.quant_builder import QuantBuilder
-from sklearn.neural_network import MLPClassifier
-from sklearn.preprocessing import StandardScaler
 
-from alphadia import fdr, grouping, libtransform, utils
+from alphadia import grouping, libtransform, utils
 from alphadia.consensus.utils import read_df, write_df
 from alphadia.constants.keys import ConfigKeys, StatOutputKeys
 from alphadia.constants.settings import FIGURES_FOLDER_NAME
-from alphadia.exceptions import NoPsmFoundError, TooFewProteinsError
-from alphadia.fdrx.utils import train_test_split_
+from alphadia.exceptions import NoPsmFoundError
 from alphadia.outputaccumulator import (
     AccumulationBroadcaster,
     TransferLearningAccumulator,
 )
+from alphadia.outputtransform_.protein_fdr import perform_protein_fdr
+from alphadia.outputtransform_.quant_builder import QuantBuilder
 from alphadia.transferlearning.train import FinetuneManager
 from alphadia.workflow import manager, peptidecentric
 from alphadia.workflow.config import Config
@@ -899,102 +897,6 @@ def _build_run_internal_df(
         logger.warning(f"Error reading timing manager for {raw_name}")
 
     return pd.DataFrame(internal_dict)
-
-
-def perform_protein_fdr(psm_df: pd.DataFrame, figure_path: str) -> pd.DataFrame:
-    """Perform protein FDR on PSM dataframe"""
-
-    protein_features = []
-    for _, group in psm_df.groupby(["pg", "decoy"]):
-        protein_features.append(
-            {
-                "pg": group["pg"].iloc[0],
-                "genes": group["genes"].iloc[0],
-                "proteins": group["proteins"].iloc[0],
-                "decoy": group["decoy"].iloc[0],
-                "count": len(group),
-                "n_precursor": len(group["precursor_idx"].unique()),
-                "n_peptides": len(group["sequence"].unique()),
-                "n_runs": len(group["run"].unique()),
-                "mean_score": group["proba"].mean(),
-                "best_score": group["proba"].min(),
-                "worst_score": group["proba"].max(),
-            }
-        )
-
-    feature_columns = [
-        "count",
-        "mean_score",
-        "n_peptides",
-        "n_precursor",
-        "n_runs",
-        "best_score",
-        "worst_score",
-    ]
-
-    protein_features = pd.DataFrame(protein_features)
-
-    X = protein_features[feature_columns].values
-    y = protein_features["decoy"].values
-
-    X_train, X_test, y_train, y_test = train_test_split_(
-        X,
-        y,
-        test_size=0.2,
-        random_state=42,
-        exception=TooFewProteinsError,
-    )
-
-    scaler = StandardScaler()
-    X_train = scaler.fit_transform(X_train)
-    X_test = scaler.transform(X_test)
-
-    clf = MLPClassifier(random_state=0).fit(X_train, y_train)
-
-    protein_features["proba"] = clf.predict_proba(scaler.transform(X))[:, 1]
-    protein_features = pd.DataFrame(protein_features)
-
-    protein_features = fdr.get_q_values(
-        protein_features,
-        score_column="proba",
-        decoy_column="decoy",
-        qval_column="pg_qval",
-    )
-
-    n_targets = (protein_features["decoy"] == 0).sum()
-    n_decoys = (protein_features["decoy"] == 1).sum()
-
-    logger.info(
-        f"Normalizing q-values using {n_targets:,} targets and {n_decoys:,} decoys"
-    )
-
-    protein_features["pg_qval"] = protein_features["pg_qval"] * n_targets / n_decoys
-
-    if figure_path is not None:
-        fdr.plot_fdr(
-            X_train,
-            X_test,
-            y_train,
-            y_test,
-            clf,
-            protein_features["pg_qval"],
-            figure_path,
-        )
-
-    return pd.concat(
-        [
-            psm_df[psm_df["decoy"] == 0].merge(
-                protein_features[protein_features["decoy"] == 0][["pg", "pg_qval"]],
-                on="pg",
-                how="left",
-            ),
-            psm_df[psm_df["decoy"] == 1].merge(
-                protein_features[protein_features["decoy"] == 1][["pg", "pg_qval"]],
-                on="pg",
-                how="left",
-            ),
-        ]
-    )
 
 
 def transfer_library_stat_df(transfer_library: SpecLibBase) -> pd.DataFrame:
